fix conflict (#69)

+++ x/oracle/keeper/owasm.go
@@ -16,7 +16,7 @@ import (
 const gasConversionFactor = 20_000_000
 
 func ConvertToOwasmGas(cosmos uint64) uint64 {
-	return uint64(cosmos * gasConversionFactor)
+	return cosmos * gasConversionFactor
 }
 
 func ConvertToGas(owasm uint64) uint64 {
